Version 2! Still with 2 levels
Updated various things here to make the game a bit more playable:

1. Moved the Score and Life counters to the top of the screen
2. Added a divider to separate the top of the screen from the playable
area. The ball now bounces off of the divider
3. Added the stage name and number to the top of the screen
4. Made "final" versions of the first two stages.
5. Changed the default background color to black.
6. Added a Text class from which the Counter class (and Score/Life) is
derived.
7. Greatly improved the realism of the ball bouncing off bricks. Now the
ball bounces off the visible pixels of bricks rather than the image
rectangle. It also bounces according to the angle of entry on the brick
surface (not just cardinal direction-based).
8. Added the ability to press "R" to reset the ball to its initial
position because the ball now sometimes gets stuck in an unrecoverable
cycle.